b27 refactoring + add count to topbar
